+++ src/LionWeb.Protocol.Delta/Message/Event/IPartitionDeltaEvent.cs
@@ -300,7 +300,7 @@ public record ChildReplaced(
     }
 }
 
-public record ChildMovedFromOtherContainment(
+public record ChildMovedFromContainmentInOtherParent(
     TargetNode NewParent,
     MetaPointer NewContainment,
     Index NewIndex,